make builds pass; rename types to schema_types
