+++ mosaic/mosaic_shared/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2026 Pasteur Labs. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from .types import (
+from .schema_types import (
     BCType,
     FaceBC,
     GridBC,
