+++ mosaic/mosaic_shared/problems/navier_stokes_grid/grid_fluid.py
@@ -5,7 +5,7 @@
 
 All functions operate on raw numpy arrays with shape convention (N, N, 1, 2)
 for vector fields and (N, N) for scalar fields, matching GridVectorField /
-GridScalarField from mosaic_shared.types.
+GridScalarField from mosaic_shared.schema_types.
 """
 
 import numpy as np
