+++ mosaic/mosaic_shared/problems/navier_stokes_grid/schemas.py
@@ -5,13 +5,13 @@
 
 The base schemas carry plain (non-`Differentiable`) array types.  Each solver
 wraps the fields it actually supports gradients on via
-``mosaic_shared.types.make_differentiable``::
+``mosaic_shared.schema_types.make_differentiable``::
 
     from mosaic_shared.problems.navier_stokes_grid import (
         InputSchema as _Base,
         OutputSchema as _BaseOut,
     )
-    from mosaic_shared.types import make_differentiable
+    from mosaic_shared.schema_types import make_differentiable
 
     InputSchema = make_differentiable(
         _Base, ["v0", "viscosity", "dt", "inflow_profile"]
@@ -23,7 +23,7 @@ Solvers with extra inputs (e.g. ``inner_steps``) should subclass the result of
 """
 
 import numpy as np
-from mosaic_shared.types import GridBC, GridObstacle, GridVectorField
+from mosaic_shared.schema_types import GridBC, GridObstacle, GridVectorField
 from pydantic import BaseModel, Field, model_validator
 from tesseract_core.runtime import Array, Float32
