+++ mosaic/mosaic_shared/problems/structural_mesh/schemas.py
@@ -13,13 +13,13 @@ Canonical interface
 
 The base schemas carry plain (non-`Differentiable`) array types.  Each solver
 wraps the fields it actually supports gradients on via
-``mosaic_shared.types.make_differentiable``::
+``mosaic_shared.schema_types.make_differentiable``::
 
     from mosaic_shared.problems.structural_mesh import (
         InputSchema as _Base,
         OutputSchema as _BaseOut,
     )
-    from mosaic_shared.types import make_differentiable
+    from mosaic_shared.schema_types import make_differentiable
 
     InputSchema = make_differentiable(_Base, ["rho"])
     OutputSchema = make_differentiable(_BaseOut, ["compliance"])
@@ -29,7 +29,7 @@ Solvers with additional material parameters (E, nu, xmin) should subclass
 """
 
 import numpy as np
-from mosaic_shared.types import HexMesh, MeshBC, MeshDirichletBC, MeshNeumannBC
+from mosaic_shared.schema_types import HexMesh, MeshBC, MeshDirichletBC, MeshNeumannBC
 from pydantic import BaseModel, Field
 from tesseract_core.runtime import Array, Float32
