+++ mosaic/mosaic_shared/schema_types.py
@@ -51,7 +51,7 @@ class GridBC(BaseModel):
 
     Example — 2-D channel flow (periodic in x, no-slip walls in y)::
 
-        from mosaic_shared.types import BCType, FaceBC, GridBC
+        from mosaic_shared.schema_types import BCType, FaceBC, GridBC
 
         bc = GridBC(
             y_lo=FaceBC(type=BCType.NO_SLIP),
@@ -180,7 +180,7 @@ class MeshBC(BaseModel):
     Example — 3-D Stokes no-slip walls::
 
         import numpy as np
-        from mosaic_shared.types import MeshBC, MeshDirichletBC
+        from mosaic_shared.schema_types import MeshBC, MeshDirichletBC
 
         bc = MeshBC(
             dirichlet=MeshDirichletBC(
