+++ mosaic/tesseracts/navier-stokes-grid/phiflow/tesseract_api.py
@@ -13,7 +13,7 @@ from mosaic_shared.problems.navier_stokes_grid import (
     OutputSchema as _CanonicalOutputSchema,
 )
 from mosaic_shared.problems.navier_stokes_grid import drag_jax
-from mosaic_shared.types import make_differentiable
+from mosaic_shared.schema_types import make_differentiable
 from phi.jax.flow import (
     Box,
     CenteredGrid,
