+++ mosaic/tesseracts/navier-stokes-grid/warp-ns/tesseract_api.py
@@ -19,7 +19,7 @@ from mosaic_shared.problems.navier_stokes_grid import (
 from mosaic_shared.problems.navier_stokes_grid import (
     OutputSchema as _CanonicalOutputSchema,
 )
-from mosaic_shared.types import make_differentiable
+from mosaic_shared.schema_types import make_differentiable
 from pydantic import Field
 
 wp.init()
