+++ mosaic/tesseracts/structural-mesh/jax-fem/tesseract_environment.yaml
@@ -19,3 +19,4 @@ dependencies:
       - pyfiglet==1.0.3
       - jax[cuda12]==0.5.3
       - jax-fem==0.0.9
+      - equinox==0.11.12
