+++ mosaic/tesseracts/structural-mesh/jax-fem/tesseract_requirements.txt
@@ -7,4 +7,5 @@ fenics-basix==0.9.0
 pyfiglet==1.0.3
 jax[cuda12]==0.5.3
 jax-fem==0.0.9
+equinox==0.11.12
 ../../../mosaic_shared
