+++ mosaic/tesseracts/thermal-mesh/torch-fem-thermal/tesseract_api.py
@@ -30,7 +30,7 @@ from mosaic_shared.problems.thermal_mesh import (
 from mosaic_shared.problems.thermal_mesh import (
     OutputSchema as _CanonicalOutputSchema,
 )
-from mosaic_shared.types import make_differentiable
+from mosaic_shared.schema_types import make_differentiable
 from pydantic import Field
 from tesseract_core.runtime import ShapeDType
 from torchfem import SolidHeat
