+++ tests/dummy_tesseracts/navier_stokes_grid/tesseract_api.py
@@ -44,7 +44,7 @@ from mosaic_shared.problems.navier_stokes_grid import (
 from mosaic_shared.problems.navier_stokes_grid import (
     OutputSchema as _CanonicalOutputSchema,
 )
-from mosaic_shared.types import make_differentiable
+from mosaic_shared.schema_types import make_differentiable
 from pydantic import ConfigDict
 from tesseract_core.runtime import ShapeDType
