+++ tests/dummy_tesseracts/thermal_mesh/tesseract_api.py
@@ -43,7 +43,7 @@ from mosaic_shared.problems.thermal_mesh import (
 from mosaic_shared.problems.thermal_mesh import (
     OutputSchema as _CanonicalOutputSchema,
 )
-from mosaic_shared.types import make_differentiable
+from mosaic_shared.schema_types import make_differentiable
 from pydantic import ConfigDict
 from tesseract_core.runtime import ShapeDType
